[neovim] extract gitsigns plugin definition and config
